refactor: extract connected components into bundle-utils, add tests and README
- Move connected component analysis (findAllReferences, findConnectedComponents,
  splitBundleByDependencies) from bundle.ts to new bundle-utils.ts
- Make low-level graph functions internal; only export splitBundleByDependencies,
  SplitBundleOptions, and redirectReferences as public API
- Fix remaining ESLint JSDoc issues (missing @param descriptions, @returns)
- Fix stray JSON.stringify -> stringify in packComponentsIntoBatches
- Add 6 new bot tests (17 total):
  - Cross-batch reference redirection (Patient ref -> conditional ref)
  - Multiple disconnected components processed independently
  - Reference to stored Bundle input
  - Multi-attachment DocumentReference -> Binary co-location
  - PractitionerRole grouped with Practitioner in same phase
  - Preferred identifier selection for conditional refs
- Add README documentation for FHIR Bundle Batching Bot
